Catchup-Mohith-main/backend/pyproject.toml
# /home/mohith/Catchup-Mohith/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stockbridge"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.29.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "redis>=5.0.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
  "temporalio>=1.4.0",
  "sendgrid>=6.11.0",
  "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.4.0",
  "black>=24.0.0",
  "psycopg2-binary>=2.9.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "N"]
ignore = ["N818"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
